flask

+++ app.py
@@ -58,17 +58,21 @@ def dbtest():
 
 @app.route('/add',methods = ["GET"])
 def add_get():
-    return render_template("add.html")
+    if 'user_id' in session:
+        return render_template("add.html")
+    else:
+        return redirect('/login')
 
 
 @app.route('/add',methods = ["POST"])
 def add_post():
+    user_id = session['user_id']
     # 入力フォームのデータを取得
     task = request.form.get("task")
     # DB接続
     conn = sqlite3.connect('dbtest.db')
     c = conn.cursor()
-    c.execute("insert into task values(null,?)",(task,))
+    c.execute("insert into task values(null,?,?)",(task,user_id))
     conn.commit()
     c.close()
     return redirect("/list")
@@ -76,33 +80,41 @@ def add_post():
 
 @app.route("/list")
 def list():
-    conn = sqlite3.connect('dbtest.db')
-    c = conn.cursor()
-    c.execute("select id,task from task")
-    task_list = []
-    for row in c.fetchall():
-        # i = row [0]
-        # print(str(i)+"週目"+str(row))
-        task_list.append({"id":row[0],"task":row[1]})
-
-    c.close()
-    # print(task_list)
-    return render_template("list.html", task_list=task_list)
+    if 'user_id' in session:
+        user_id = session['user_id']
+        conn = sqlite3.connect('dbtest.db')
+        c = conn.cursor()
+        c.execute("select name from user where id = ?",(user_id,))
+        user_name = c.fetchone()[0]
+        c.execute("select id,task from task where user_id = ?",(user_id,))
+        task_list = []
+        for row in c.fetchall():
+            # i = row [0]
+            # print(str(i)+"週目"+str(row))
+            task_list.append({"id":row[0],"task":row[1]})
+        c.close()
+        # print(task_list)
+        return render_template("list.html", task_list=task_list , user_name = user_name)
+    else:
+        return redirect('/login')
 
 
 @app.route('/edit/<int:id>')
 def edit(id):
-    conn = sqlite3.connect('dbtest.db')
-    c = conn.cursor()
-    c.execute("select task from task where id = ?",(id,))
-    task = c.fetchone()
-    c.close()
-    if task is not None:
-        task = task[0]
+    if 'user_id' in session:
+        conn = sqlite3.connect('dbtest.db')
+        c = conn.cursor()
+        c.execute("select task from task where id = ?",(id,))
+        task = c.fetchone()
+        c.close()
+        if task is not None:
+            task = task[0]
+        else:
+            return "タスクがないよ"
+        item = {"id":id, "task":task}
+        return render_template("edit.html",item=item)
     else:
-        return "タスクがないよ"
-    item = {"id":id, "task":task}
-    return render_template("edit.html",item=item)
+        return redirect('/login')
 
 
 @app.route('/edit',methods=['POST'])
@@ -130,19 +142,25 @@ def edit_post():
 
 @app.route("/del/<int:id>")
 def del_task(id):
-    # データベースと接続
-    conn = sqlite3.connect('dbtest.db')
-    c = conn.cursor()
-    # データの更新
-    c.execute("delete from task where id = ?",(id,))
-    conn.commit()
-    c.close()
-    return redirect("/list")
+    if 'user_id' in session:
+        # データベースと接続
+        conn = sqlite3.connect('dbtest.db')
+        c = conn.cursor()
+        # データの更新
+        c.execute("delete from task where id = ?",(id,))
+        conn.commit()
+        c.close()
+        return redirect("/list")
+    else:
+        return redirect('/login')
 
 
 @app.route('/regist', methods=['GET'])
 def regist_get():
-    return render_template("regist.html")
+    if 'user_id' in session:
+        return redirect('/list')
+    else:
+        return render_template("regist.html")
 
 
 @app.route('/regist',methods=['POST'])
@@ -156,12 +174,15 @@ def regist_post():
     c.execute("insert into user values(null,?,?)",(name,password))
     conn.commit()
     c.close()
-    return "登録完了"
+    return redirect('/list')
 
 
 @app.route('/login',methods=['GET'])
 def login_get():
-    return render_template('login.html')
+    if 'user_id' in session:
+        return redirect('/login')
+    else:
+        return render_template('login.html')
 
 
 @app.route('/login',methods=['POST'])
@@ -180,6 +201,12 @@ def login_post():
     else:
         session['user_id'] =  user_id[0]
     return redirect('/list')
+
+
+@app.route('/logout')
+def logout():
+    session.pop('user_id',None)
+    return redirect('/login')
